chore: add awq support
